Split `PurchasesTests` step 5: extracted `PurchasesGetOfferingsTests`
Follow up to #1609.
